backend/requirements.txt
# Existing requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
groq>=0.13.0
httpx==0.27.2
aiohttp==3.9.1

# RAG-specific requirements
chromadb==0.4.18
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
pypdf==3.17.4
beautifulsoup4==4.12.2
requests==2.31.0
python-multipart==0.0.6  # For file uploads
tiktoken==0.5.2  # Token counting

# pip install sentence-transformers==2.3.1
# pip install huggingface-hub==0.20.3